Move inputs and solution to data directory
I discovered recently the request by the AoC author to not share inputs
publicly. In order to support that more easily, I'm separating the
puzzle input and the solutions for the two parts of each puzzle to their
own directory (data) so that I can more easily ignore them in git.

The daily tests will now compare results against two fixtures that have
already read the first line of the part1.solution and part2.solution
files, respectively.

This has the side effect of making it easier to write solutions in other
languages and check results outside of the Python tests.